Update 16/05/2021 - The start
Created the project, with an empty scene. I made a platform and started work on a character controller. I'm not super familiar with Unity's 2D physics engine as opposed to its 3D one, but I made something that moves pleasingly left and right, when inputs are made. Jumping doesn't work for some reason, but it's not bad for roughly fifteen minutes of work.