Update 24/05/2021: Attempted to make Unity's physics system actually work in the way that it's supposed to
I attempted to make the jumping code work again, and also experimented with making a raycast check to ensure the player is grounded before being able to jump.

However, this code simply does not work for some reason. Despite the if statement running when the jump button is pressed (there's a debug message that gets printed), the command that adds force simply doesn't run for some reason. Additionally, the grounded check bool simply does not run, not even displaying the console message that should run every time the bool is called, regardless of if it returns true or false. This is something I need to figure out next time, as it's late and I need to go to bed.

 Despite being called