Update 27/07/2021 - Fixing friction and attempting to fix grounding detection
To fix the problem of the player moving weirdly slowly on the ground, I created a new Physics Material 2D, with the friction value set to zero. I might need to increase this slightly so the player doesn’t slide forever when hit by a friction force.

I also attempted to fix my grounding detection code. I realised I was using a regular Raycast, when I should’ve been using Raycast2D. This worked – too well. Now instead of never registering, it always registers even when the player is blatantly airborne. This may be because of some kind of glitch involving the raycast hitting the player’s collider, which is a possibility I heard on a forum, although I have not fully tested it yet due to having other obligations to work on today.

I also added some extra platforms to the scene, instead of having just a single platform to walk around on.