Update 17/08/2021 - Aiming controls and work began on tilemaps, crouching and a gun script
I made great progress on the arduous process of learning Unity's 2D Tilemap system. While I'm sure it will be quite versatile to use in the long run, the process of actually setting it up is extremely aggravating. This is primarily due to finicky and unintuitive controls, where almost every action you want to perform seems to take twice as many button presses as you’d expect it to, you’re never sure what updating a transform value on a tile palette will actually do, and actions seemingly undo themselves without your input. That being said, these should be part of the learning curve, and I anticipate I will learn the tricks required to use the system efficiently.

I also created some code for letting the player ‘aim’ by pressing one of two buttons to rotate their aim clockwise or counter-clockwise. The player’s aim direction is currently represented by a debug line, as there is presently no actual reticle/beam system. I am presently pleased with the responsiveness of the aim controls, but I will need to sort out a suitable control scheme for it, and there is presently nothing that needs to be aimed. This is something I made a token amount, by creating a ‘Gun’ class. This class presently does nothing except contain some variables, since I realised there were far more important things to work on, and proceeded to work on adding a Tilemap like I mentioned above.

I also started work on a function to let the player crouch, but it is untested and there is presently no control for it.