Aim controller now uses mouse cursor rather than rotating on axis
I realised the old aiming code (which used two keyboard buttons to rotate the aim axis) was incredibly awkward and uncomfortable to use. I reworked it to now use the mouse cursor, so the player will aim in the direction of whatever the mouse is hovering over. Simple and effective!