Reworked movement controller to no longer use Rigidbody2D.MovePosition()
When run, this function will pause all physics operations on the current frame. This prevents other physics effects on the player from functioning normally, most notably that gravity will not trigger properly, causing an incredibly slow fall. To fix this, I devised a slightly awkward but functional system that sets the player’s current rigidbody velocity, but checks to make sure the current velocity is not greater than the desired one. This will steer the player’s movement velocity towards the desired direction, without overshooting and causing the player to accelerate beyond the desired speed.