ResourceMeter base class and SegmentedMeter inherited class
Created a base ResourceMeter class to handle different kinds of HUD meters. This meter is set up and updated by referencing a Resource struct.
I then created an inherited class, SegmentedMeter, for resource meters consisting of a quantity of objects that are depleted one by one, such as ammunition in a gun. This consists of code to generate a series of images, and colour them based on the remaining quantity in the Resource and its critical status.