Implemented gun firing code with ammo depletion
Added ‘capacity’ Resource struct to determine remaining ammunition in the gun
Added ‘meter’ ResourceMeter that updates to show the remaining ammunition. The meter is attached to the gun by default, but when picked up by a player it will be auto-parented to the player’s aim socket position, allowing the player’s ammunition meter to change seamlessly based on the equipped weapon.
Added Shoot() function that checks the firing delay and remaining ammunition to ensure it can shoot, then executes the necessary firing functions. Support exists for multiple projectiles and calculating firing lines, but currently all that happens is a debug line is drawn to the correct shot position.
Started implementing Pickup() and Drop() functions to attach the weapon to and detach it from a player, allowing them to pick up different weapons. This function is not fully implemented yet, but is currently used to set up weapon attributes.
Implemented code in AimController to reference a currently equipped gun and receive input for firing it.
Even though the gun does not currently shoot anything, I am still very pleased with how this code is turning out.