Tweaked animation controller for landing animations and fixing sound bugs
I had to restructure the walking blend tree in the animation controller, as it would glitch out and cause multiple sound effects to play over each other.

I also added the new state for landing from falls, as a blend tree  that switches between the left and right variants depending on the player's look direction.