Player reticle and aiming beam
When the player moves their cursor around, a reticle will appear, with a line connecting it to the player.
Created a reticle sprite
Created a reticle Image object in the player’s HUD canvas. This is moved around using code in the AimController class, which uses ScreenToWorldPoint() to set its position to wherever the reticle is on screen. Pretty simple stuff.
If the reticle is a certain distance away from the player, a line renderer will enable, and draw a simple beam between the reticle and the weapon’s aim origin, to communicate positions.