Modified guns, aim axis and AimController rotation to not rely on 3D space code in a 2D scene setup
This was causing some confusing problems, most notably when having weapons transition between being held and being dropped items in the world. Now the guns and aim axis rely solely on the X and Y axes for position and the Z axis for rotation, just like everything else.

I did however have to rework the code for ensuring the player's gun rotates properly when their aim rotates more than 90 degrees from where it started. I added some simple code that sets the gun's X rotation to 180 if this is the case, and 0 at all other times. Additional code is present to reset the X and Y rotation values when the gun is dropped.

This change works very well.