Cosmetic effects for gun – firing noise, shell ejection and bullet impacts
I added some simple effects to improve the gun’s appeal. I added an AudioSource that is referenced to run PlayOneShot() with a firing noise whenever the gun fires, and a particle system that plays to eject a shell casing. I also added a particle system prefab that simulates debris ricocheting off a surface from a gunshot, that is spawned and activated using HitscanShot’s SpawnDebris() function wherever the player shoots a surface.