Modular CameraFollow script
This MonoBehaviour class attaches to an object with a Camera component, and runs code to make it smoothly shift its position and rotation to match a target transform, with a distance offset applied. This acts as a handy tool for smooth camera following movement for purposes such as character controllers, while still keeping the camera as a child of the object being moved around for easy organisation.