Created input actions for new input system
I've come back to this project after a while, and realised that my existing character controller code is very badly made. I plan on overhauling it, and one change I want to make is to have the controller use the new input system. This code will make it much easier to rebind controls and implement complex commands in a more modular manner.