Added variable jump height to player movement controller
If the player presses the jump button for less than the duration specified by 'jumpHoldTime', the overall jump force will be reduced, to allow the player to more precisely jump onto low platforms when a roof or hazard is present above them.

The mechanic needs some polish due to being slightly abrupt and awkward, but is still a great addition.