Experimental 'ModifyVelocity' function
This function is meant to alter the player's velocity in such a way that it doesn't override other, greater forces acting on it. This is so if the player is knocked in a direction by a larger force, they cannot immediately negate it by making any movement.

I have not properly tested this function yet.